update javascript library
